Resolve S13P11A601-47 "Feature/"

See merge request s13-webmobile1-sub1/S13P11A601!127

+++ BE/src/main/java/com/codaily/project/controller/SpecificationController.java
@@ -149,7 +149,7 @@ public class SpecificationController {
     public ResponseEntity<SpecificationTimeResponse> getTotalEstimatedTime(
             @PathVariable Long specId
     ) {
-        int totalTime = featureItemService.calculateTotalEstimatedTime(specId);
+        Double totalTime = featureItemService.calculateTotalEstimatedTime(specId);
         SpecificationTimeResponse response = new SpecificationTimeResponse(specId, totalTime);
         return ResponseEntity.ok(response);
     }
